Address PR review round 2: null safety, defensive guards, validation fixes
- RS_TileExplode: return Iterator.empty when raster is null instead of NPE
- HadoopImageInputStream: guard against non-progressing stream (ret_len == 0)
- HadoopImageInputStreamTest: fix UnstableInputStream to return at least 1 byte
- RasterOptions: only enforce tileWidth/tileHeight pairing when retile=true
- RasterPartitionReader: use floating-point division for aspect ratio check
- rasterIOTest: remove explain(true) noise from CI logs

+++ spark/common/src/main/scala/org/apache/spark/sql/sedona_sql/io/raster/RasterPartitionReader.scala
@@ -175,7 +175,7 @@ object RasterPartitionReader {
             throw new IllegalArgumentException(
               s"Internal tile size of $path contains zero ($tw x $th). " + tileSizeError)
           }
-          if (tw / th > 10 || th / tw > 10) {
+          if (tw.toDouble / th > 10.0 || th.toDouble / tw > 10.0) {
             throw new IllegalArgumentException(
               s"Internal tile shape of $path is too thin ($tw x $th). " + tileSizeError)
           }
